Fixed recently introduced compilation warnings
src\v_video.cpp(1771): warning C4244: '=': conversion from 'double' to 'int', possible loss of data
src\v_video.cpp(1773): warning C4244: '=': conversion from 'double' to 'int', possible loss of data

+++ src/v_video.cpp
@@ -1768,9 +1768,9 @@ void ScaleWithAspect (int &w, int &h, int Width, int Height)
 	}
 	double y = w/yratio;
 	if (y > h)
-		w = h*yratio;
+		w = static_cast<int>(h * yratio);
 	else
-		h = y;
+		h = static_cast<int>(y);
 }
 
 void IVideo::DumpAdapters ()
